🔧  Add commands to build the site for deployment.
To prepare for use with Netlify, the repository
is being setup to mimic being built as if it was
being compiled from source to a packaged version.

For now, the project is barebones and will just
copy the entire src/ folder into dist/ which will
be the publish directory for Netlify to use.

In the future, the project will either be built
using webpack or a static site generator (such as
Gatsby).

- Add command to clean dist/ and then copy src/.